move to new FreeCAD workbench structure
