+++ freecad/cadenas3dfindit/init_gui.py
@@ -27,7 +27,7 @@ import FreeCADGui
 
 class CADENAS3DFinditShowCommand:
     def GetResources(self):
-        import CADENAS3DfinditDummy
+        from freecad.cadenas3dfindit import CADENAS3DfinditDummy
         module_path = os.path.split(CADENAS3DfinditDummy.__file__)[0]
         return { 
           'Pixmap': os.path.join(module_path, "Resources", "icons", "3dfindit.svg"),
@@ -38,13 +38,13 @@ class CADENAS3DFinditShowCommand:
         return True
 
     def Activated(self):
-        import CADENAS3DfinditDialog
+        from freecad.cadenas3dfindit import CADENAS3DfinditDialog
         CADENAS3DfinditDialog.toggle()
 
 
 class CADENAS3DFinditGeoSearchCommand:
     def GetResources(self):
-        import CADENAS3DfinditDummy
+        from freecad.cadenas3dfindit import CADENAS3DfinditDummy
         module_path = os.path.split(CADENAS3DfinditDummy.__file__)[0]
         return { 
           'Pixmap': os.path.join(module_path, "Resources", "icons", "geomsearch.svg"),
@@ -56,7 +56,7 @@ class CADENAS3DFinditGeoSearchCommand:
 
     def Activated(self):
         # Make sure our dialog is visible.
-        import CADENAS3DfinditDialog
+        from freecad.cadenas3dfindit import CADENAS3DfinditDialog
         CADENAS3DfinditDialog.show()
 
         # Export currently active document to STP.
@@ -81,7 +81,7 @@ class CADENAS3DFinditGeoSearchCommand:
           return False
 
         # Init geometrical search.
-        import Browser
+        from freecad.cadenas3dfindit import Browser
         api = Browser.getInstance().getThreeDAPI()
         api.startGeoSearch(os.path.basename(tmpSTPPath))
 
@@ -101,9 +101,9 @@ class CADENAS3DFinditGeoSearchCommand:
 
 class CADENAS3DFinditSketchSearchCommand:
     def GetResources(self):
-        import CADENAS3DfinditDummy
+        from freecad.cadenas3dfindit import CADENAS3DfinditDummy
         module_path = os.path.split(CADENAS3DfinditDummy.__file__)[0]
-        return {
+       return {
           'Pixmap': os.path.join(module_path, "Resources", "icons", "sketchsearch.svg"),
           'MenuText': "Sketch search",
           'ToolTip': "Start a sketch search on 3DfindIT.com"}
@@ -115,12 +115,12 @@ class CADENAS3DFinditSketchSearchCommand:
         return True
 
 
-class CADENAS3DfinditWorkbench(Workbench):
+class CADENAS3DfinditWorkbench(FreeCADGui.Workbench):
     def __init__(self):
-        import CADENAS3DfinditDummy
+        from freecad.cadenas3dfindit import CADENAS3DfinditDummy
         module_path = os.path.split(CADENAS3DfinditDummy.__file__)[0]
         self.__class__.Icon = os.path.join(module_path, "Resources", "icons", "3dfindit.svg")
-        self.__class__.MenuText = "3DfindIT.com"
+       self.__class__.MenuText = "3DfindIT.com"
         self.__class__.ToolTip = "3DfindIT.com by CADENAS"
 
     def Initialize(self):
@@ -140,7 +140,7 @@ class CADENAS3DfinditWorkbench(Workbench):
     def GetClassName(self): 
         return "Gui::PythonWorkbench"
 
-freeCadVersion = int(App.Version()[1])
+freeCadVersion = int(FreeCAD.Version()[1])
 pythonVersion = int(sys.version[0:1])
 if freeCadVersion >= 19 and pythonVersion >= 3:
     FreeCADGui.addWorkbench(CADENAS3DfinditWorkbench())
